third_party/blink/web_tests/platform/generic/fast/css/getComputedStyle/computed-style-recalc-expected.txt

Check that getComputedStyle causes a style recalc when necessary, and only when necessary.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS successfullyParsed is true

TEST COMPLETE
